dafsa: widen sequence counts/ids to u64 (n>=17 exceeds u32::MAX)
ZZ12 cumulative rat counts cross u32::MAX (4.29e9) around n=17
(~12.3e9). The block wire format already stored per-state rank
counts as u64, and state/edge ids stay u32 (only ~25M states at
n=17), so this is purely a widening of sequence-space quantities:
no block-format change, no storage growth, no version bump (still
test phase).

Widened:
- BlockManifest.n_sequences u32 to u64; len() returns u64 (both
  readers).
- index_of returns Option<u64>; rank accumulator is u64 and drops
  the "sub_rec.count as u32" truncation (the one active bug: a
  sibling class above u32::MAX silently wrapped the returned id).
- get(i) takes u64 (was usize, which is 32-bit on wasm32, the
  deploy target); the internal "remaining" was already u64.
- RatDafsa::index_of (in-memory) drops its "as u32" cast.
- AssetParams.n_sequences and the rat_enum writer casts to u64.
- WASM boundary db_id_of/db_seq_of use f64 (exact JS number for
  every integer below 2^53), matching db_init's existing "total",
  so the JS explorer keeps doing ordinary number arithmetic.

Tests: a manifest with n_sequences above u32::MAX round-trips and
len() reports it; a forged block where an earlier sibling class of
5e9 makes index_of return a rank above u32::MAX (truncated to
705032704 before the fix). wasm32 builds clean.

Also: stop tracking a stray __pycache__ .pyc and gitignore them.

+++ src/stringmatch/dafsa/rat.rs
@@ -179,11 +179,11 @@ impl RatDafsa {
 
     /// Assigned external index of `rat` (its position in
     /// `(length, lex)` order) if it is in the set, else `None`.
-    pub fn index_of<S: AsRef<[i8]>>(&self, rat: S) -> Option<u32> {
+    pub fn index_of<S: AsRef<[i8]>>(&self, rat: S) -> Option<u64> {
         let prefixed = prefix(rat.as_ref());
         // Lex rank in the inner DAFSA == (length, lex) rank of the
         // raw rat -- exactly the external assigned index.
-        self.inner.lex_rank_of(&prefixed).map(|r| r as u32)
+        self.inner.lex_rank_of(&prefixed).map(|r| r as u64)
     }
 
     /// Set membership: does `rat` appear in this `RatDafsa`?
@@ -366,7 +366,7 @@ mod tests {
 
         // index_of returns the (length, lex) position.
         for (i, e) in expected.iter().enumerate() {
-            assert_eq!(rd.index_of(e.as_slice()), Some(i as u32));
+            assert_eq!(rd.index_of(e.as_slice()), Some(i as u64));
         }
         assert_eq!(rd.index_of([42i8, 42].as_slice()), None);
 
@@ -386,7 +386,7 @@ mod tests {
         assert_eq!(restored_iter, expected);
         for (i, e) in expected.iter().enumerate() {
             assert_eq!(restored.get(i).as_ref(), Some(e));
-            assert_eq!(restored.index_of(e.as_slice()), Some(i as u32));
+            assert_eq!(restored.index_of(e.as_slice()), Some(i as u64));
         }
     }
